fix: preserve inline code spans in llms text generation
Inline code such as `<ArdoRoot>` in prose collapsed to an empty span because the component-tag scanner jumped to the next `>` and consumed the closing backtick. Skip backtick-delimited spans verbatim, matching the existing code-fence handling. The MDX stripping logic moves to a dedicated llms-mdx-strip.ts module to keep both files within the max-lines limit.
